+++ .travis/install.sh
@@ -1,0 +1,6 @@
+#!/bin/sh
+
+set -e
+set -u
+
+pip install tox;