+++ timezonefinder/helpers_numba.py
@@ -1,11 +1,10 @@
 # -*- coding:utf-8 -*-
 from math import asin, atan2, ceil, cos, degrees, floor, radians, sin, sqrt
 
+from numba import b1, f8, i2, i4, njit, typeof, u2, u8
 from numpy import int64
 
-from numba import b1, f8, i2, i4, njit, typeof, u2, u8
-
-from .global_settings import COORD2INT_FACTOR, INT2COORD_FACTOR, MAX_HAVERSINE_DISTANCE
+from timezonefinder.global_settings import COORD2INT_FACTOR, INT2COORD_FACTOR, MAX_HAVERSINE_DISTANCE
 
 # from global_settings import COORD2INT_FACTOR, INT2COORD_FACTOR, MAX_HAVERSINE_DISTANCE